+++ src/GridCal/Engine/Core/DataStructures/circuit_to_data.py
@@ -547,6 +547,7 @@ def get_branch_data(circuit: MultiCircuit, bus_dict, Vbus, apply_temperature,
         if time_series:
             data.branch_active[i, :] = elm.active_prof
             data.branch_rates[i, :] = elm.rate_prof
+            data.branch_contingency_rates[i, :] = elm.rate_prof * elm.contingency_factor
 
             if opf:
                 data.branch_cost[i, :] = elm.Cost_prof
@@ -554,6 +555,7 @@ def get_branch_data(circuit: MultiCircuit, bus_dict, Vbus, apply_temperature,
         else:
             data.branch_active[i] = elm.active
             data.branch_rates[i] = elm.rate
+            data.branch_contingency_rates[i] = elm.rate * elm.contingency_factor
 
             if opf:
                 data.branch_cost[i] = elm.Cost