+++ src/GridCal/Engine/Devices/dc_line.py
@@ -31,7 +31,8 @@ class DcLine(EditableDevice):
                  rate=1.0, active=True, tolerance=0, cost=0.0,
                  mttf=0, mttr=0, r_fault=0.0, x_fault=0.0, fault_pos=0.5,
                  length=1, temp_base=20, temp_oper=20, alpha=0.00330,
-                 template=None, rate_prof=None, Cost_prof=None, active_prof=None, temp_oper_prof=None):
+                 template=None, rate_prof=None, Cost_prof=None, active_prof=None, temp_oper_prof=None,
+                 contingency_factor=1.0):
         """
 
         :param bus_from:
@@ -72,7 +73,12 @@ class DcLine(EditableDevice):
                                                   'bus_to': GCProp('', DeviceType.BusDevice,
                                                                    'Name of the bus at the "to" side of the line.'),
                                                   'active': GCProp('', bool, 'Is the line active?'),
+
                                                   'rate': GCProp('MW', float, 'Thermal rating power of the line.'),
+
+                                                  'contingency_factor': GCProp('p.u.', float,
+                                                                               'Rating multiplier for contingencies.'),
+
                                                   'mttf': GCProp('h', float, 'Mean time to failure, '
                                                                  'used in reliability studies.'),
                                                   'mttr': GCProp('h', float, 'Mean time to recovery, '
@@ -156,7 +162,7 @@ class DcLine(EditableDevice):
 
         # line rating in MW
         self.rate = rate
-
+        self.contingency_factor = contingency_factor
         self.rate_prof = rate_prof
 
         # branch type: Line, Transformer, etc...