+++ src/GridCal/Engine/Simulations/ContingencyAnalysis/contingency_analysis_ts_driver.py
@@ -153,7 +153,7 @@ class ContingencyAnalysisTimeSeries(QThread):
         self.progress_text.emit('Computing branch base flows...')
         Pbus = ts_numeric_circuit.Sbus.real
         flows = linear_analysis.get_flows_time_series(Pbus)
-        rates = ts_numeric_circuit.Rates.T
+        rates = ts_numeric_circuit.ContingencyRates.T
 
         self.progress_text.emit('Computing N-1 flows...')