+++ doc/change_log.rst
@@ -10,6 +10,13 @@ This section describes the changes introduced at each Version.
 this is because the stupid policy of pypi to not allow to correct packages.
 Hence if something goes wrong, you need to re-upload with a new Version number.
 
+Version 4.1.2
+^^^^^^^^^^^^^^^
+
+- Added branch contingency multiplier for contingency related simulations
+- Added Available Transmission Capacity
+
+
 Version 4.1.1
 ^^^^^^^^^^^^^^^